requirements.txt
Django==4.2.6
djangorestframework==3.14.0
Pillow==10.0.1
django-filter==23.3
flake8==6.1.0